fix(spec): a round's tokens are the ones a round produced
The four sidecar loops argmax a bonus token out of the prefill forward
and emit it before the round loop starts; the two-model loops emit
nothing outside their loop. Dividing `emitted` by `rounds` therefore read
+1/rounds high on four of six loops — about 3% at a 128-token bench — and
made the one figure the record exists to compare incomparable between
them. The identity `1 + accept_rate * (block - 1)` was false for those
four even at a fixed block, which is what the docs claimed it was for.

`SpecLoop::seed_tokens` is what a loop emits outside its rounds, it is
subtracted, and it is on the `done` line so a reader aggregating several
requests can subtract it too.

The fixture modelled the two-model accounting under a sidecar loop's
name, so the off-by-one could not appear in the only test of that
identity; it now builds its emission count from the loop it names and the
identity is asserted on all six. Its three timings were 200/600/1000 ms,
which made draft-per-round and loop-per-round both 25 — a swap of the two
derivations passed every assertion. They are now pairwise distinct and a
test says why.

A phase timer reaching outside its round loop makes the residual
negative, which ingest refuses while naming no field. The value stays
visible and the violation is named. The `done` line also carries
`loop_kind`, because six callsites became one and the target no longer
distinguishes them.

+++ crates/rmlx-models/src/speculative/eagle3/mod.rs
@@ -947,7 +947,7 @@ pub fn eagle3_generate_greedy(
         // happened, so it still leaves exactly one record.
         super::RoundStats {
             loop_kind: super::SpecLoop::Eagle3,
-            block_size: block_total,
+            block_size: block_total as u32,
             rounds: 0,
             emitted: emitted.len(),
             total_draft: 0,
@@ -1208,7 +1208,7 @@ pub fn eagle3_generate_greedy(
     let round_loop_ns = round_loop_t0.elapsed().as_nanos();
     super::RoundStats {
         loop_kind: super::SpecLoop::Eagle3,
-        block_size: block_total,
+        block_size: block_total as u32,
         rounds,
         emitted: emitted.len(),
         total_draft,
