+++ crates/rmlx-models/src/speculative/mtp.rs
@@ -768,7 +768,7 @@ pub fn mtp_generate_greedy(
         // happened, so it still leaves exactly one record.
         super::RoundStats {
             loop_kind: super::SpecLoop::MtpSidecar,
-            block_size: block_total,
+            block_size: block_total as u32,
             rounds: 0,
             emitted: emitted.len(),
             total_draft: 0,
@@ -913,7 +913,7 @@ pub fn mtp_generate_greedy(
     let round_loop_ns = round_loop_t0.elapsed().as_nanos();
     super::RoundStats {
         loop_kind: super::SpecLoop::MtpSidecar,
-        block_size: block_total,
+        block_size: block_total as u32,
         rounds,
         emitted: emitted.len(),
         total_draft,
